Complete header
Added the header content and styled it as per the task. I contained the
entire header in another div to give it background colour so it could
abide by margin without affecting the colour going to page edge.

I haven't actually supplied a 'link' to the three header links but it
wouldn't be difficult to do so. I have already applied the decoration
and list-style properties that would style the links would they be real.