Complete first section
Breaking down the page into six sections. Second section complete. It
isn't exactly like the image but I think it is because my browser that I
use to check progress gives me a larger view window, so small details
like how much the text is wrapped etc aren't exact. But the styling and
position of elements is there.