Complete section 5
Although I didn't need to give the background a colour. I did add a
container because I think I'll need it when I get to the footer. I have
a suspicion that the footer won't stick to the bottom of the page as I
currently have it set up. But we will see when it's finished.